chore(RoundParser): relocate from domain to parser package to match responsibility
